Bugfix: Fix error where user verification page got stuck when logging in
