+++ app/(routes)/my_profile/page.tsx
@@ -1,10 +1,6 @@
 import TryAgainForm from "@/app/(routes)/my_profile/try-again-form";
-import { Skeleton } from "@/app/components/ui/skeleton";
-import CompleteProfileModal from "@/app/components/user_profile/complete-profile-modal";
-import PrivateProfileOverview from "@/app/components/user_profile/private_profile/overview";
-import PublicProfile from "@/app/components/user_profile/public_profile/profile";
-import UserProfileBanner from "@/app/components/users/user-profile-banner";
-import { fetchSubcategories } from "@/app/lib/subcategories/actions";
+import UserProfile from "@/app/components/pages/user-profile";
+import { UserProfileSkeleton } from "@/app/components/user_profile/skeleton";
 import {
 	cachedFetchUserProfileByClerkId,
 	getCurrentClerkUser,
@@ -42,21 +38,9 @@ export default async function Page() {
 		);
 	}
 
-	const subcategories = fetchSubcategories();
-
 	return (
-		<div className="container p-3 md:p-6 flex flex-col gap-2">
-			<Suspense fallback={<Skeleton className="h-20 w-full" />}>
-				<UserProfileBanner profile={profile} />
-			</Suspense>
-			<div className="grid grid-cols-1 md:grid-cols-2 gap-2">
-				<PublicProfile profile={profile} />
-				<PrivateProfileOverview profile={profile} />
-			</div>
-			<CompleteProfileModal
-				subcategoriesPromise={subcategories}
-				profile={profile}
-			/>
-		</div>
+		<Suspense fallback={<UserProfileSkeleton />}>
+			<UserProfile profile={profile} />
+		</Suspense>
 	);
 }
